mux: Extract mux-related features from `avocado.core.tree`
The `avocado.core.tree` contains multiplex/mux-related features like
multiplex and ctrl, which are not really useful for basic params. Let's
simplify the base tree object and put the extracted features in
`avocado.core.mux`.

This is a first step and the `avocado.core.mux` should contain most of
the generic parts useful for creating varianter plugins.
