+++ avocado/core/mux.py
@@ -1,0 +1,101 @@
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 2 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+#
+# See LICENSE for more details.
+#
+# Copyright: Red Hat Inc. 2016
+#
+# Authors: Lukas Doktor <ldoktor@redhat.com>
+
+"""
+This file contains mux-enabled implementations of parts useful for creating
+a custom Varianter plugin.
+"""
+import re
+
+from . import tree
+
+
+#
+# Multiplex-enabled tree objects
+#
+REMOVE_NODE = 0
+REMOVE_VALUE = 1
+
+
+class Control(object):  # Few methods pylint: disable=R0903
+
+    """ Container used to identify node vs. control sequence """
+
+    def __init__(self, code, value=None):
+        self.code = code
+        self.value = value
+
+
+class MuxTreeNode(tree.TreeNode):
+
+    """
+    Class for bounding nodes into tree-structure with support for
+    multiplexation
+    """
+
+    def __init__(self, name='', value=None, parent=None, children=None):
+        super(MuxTreeNode, self).__init__(name, value, parent, children)
+        self.ctrl = []
+        self.multiplex = None
+
+    def __repr__(self):
+        return 'TreeNode(name=%r)' % self.name
+
+    def merge(self, other):
+        """
+        Merges `other` node into this one without checking the name of the
+        other node. New values are appended, existing values overwritten
+        and unaffected ones are kept. Then all other node children are
+        added as children (recursively they get either appended at the end
+        or merged into existing node in the previous position.
+        """
+        for ctrl in other.ctrl:
+            if isinstance(ctrl, Control):
+                if ctrl.code == REMOVE_NODE:
+                    remove = []
+                    regexp = re.compile(ctrl.value)
+                    for child in self.children:
+                        if regexp.match(child.name):
+                            remove.append(child)
+                    for child in remove:
+                        self.children.remove(child)
+                elif ctrl.code == REMOVE_VALUE:
+                    remove = []
+                    regexp = re.compile(ctrl.value)
+                    for key in self.value.iterkeys():
+                        if regexp.match(key):
+                            remove.append(key)
+                    for key in remove:
+                        self.value.pop(key, None)
+        super(MuxTreeNode, self).merge(other)
+        if other.multiplex is True:
+            self.multiplex = True
+        elif other.multiplex is False:
+            self.multiplex = False
+
+
+class MuxTreeNodeDebug(MuxTreeNode, tree.TreeNodeDebug):
+
+    """
+    Debug version of TreeNodeDebug
+    :warning: Origin of the value is appended to all values thus it's not
+    suitable for running tests.
+    """
+
+    def __init__(self, name='', value=None, parent=None, children=None,
+                 srcyaml=None):
+        MuxTreeNode.__init__(self, name, value, parent, children)
+        tree.TreeNodeDebug.__init__(self, name, value, parent, children,
+                                    srcyaml)
